% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Box.R
\name{Box}
\alias{Box}
\title{Box}
\usage{
Box(text = NULL, horizontal_align = "left", vertical_align = "top",
  wrap_text = TRUE, background_color = "Transparent",
  font_family = "sans-serif", font_size = 11, font_color = "black",
  font_bold = FALSE, font_italic = FALSE, font_underline = FALSE,
  font_strikethrough = FALSE, border_width = 0,
  border_color = "Transparent", border_style = "Solid", as_html = FALSE,
  width = NULL, height = NULL)
}
\arguments{
\item{text}{character. Adding text to the box.}

\item{horizontal_align}{horizontal alignment of \code{text}. Possible values are c("left","center","right"). Defaults to \code{"left"}.}

\item{vertical_align}{vertical alignment of \code{text}. Possible values are \code{c("top","middle","bottom")}. Defaults to \code{"top"}.}

\item{wrap_text}{logical. Whether text wrapping is enabled. Defaults to \code{TRUE}.}

\item{background_color}{HTML color string. Sets the background color of the box. Defaults to \code{"Transparent"}.}

\item{font_family}{HTML font family string. Sets the font family of \code{text}. Defaults to \code{"sans-serif"}.}

\item{font_size}{integer. Sets the font size of \code{text} in pixels. Defaults to 11.}

\item{font_color}{HTML color string. Sets the font color of \code{text}. Defaults to \code{"black"}.}

\item{font_bold}{logical. Whether to bold \code{text}. Defaults to \code{FALSE}.}

\item{font_italic}{logical. Whether to have italic \code{text}. Defaults to \code{FALSE}.}

\item{font_underline}{logical. Whether to underline \code{text}. Defaults to \code{FALSE}.}

\item{font_strikethrough}{logical. Whether to draw a horizontal line across \code{text}. Defaults to code{FALSE}.}

\item{border_width}{integer. Sets the border width in pixels. Defaults to 0.}

\item{border_color}{HTML color string. Sets the border color. Defaults to "Transparent".}

\item{border_style}{HTML border style string. Sets border style. Defaults to "Solid".}

\item{as_html}{logical. Whether to interpret \code{text} as raw html. Defaults to \code{FALSE}.}

\item{width}{integer. Width of the widget in pixels. Defaults to NULL, which automatically calculates the width based on window size.}

\item{height}{integer. Height of the widget in pixels. Defaults to NULL, which automatically calculates the height based on window size.}
}
\description{
Creates an html Div container box with multiple features.
}

