% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhtmlMetro.R
\name{UpDown}
\alias{UpDown}
\title{UpDown}
\usage{
UpDown(arrow.style = "triangle", arrow.up = TRUE, arrow.size = 15,
  arrow.color = "black", arrow.width = 0.5, x = NULL,
  x.font.family = "sans-serif", x.font.size = 20, x.font.color = "black",
  x.width = 0.5, unit.width = 0.3, unit.height = 0.5, topunit = NULL,
  topunit.font.family = "sans-serif", topunit.font.color = "black",
  topunit.font.size = 12, bottomunit = NULL,
  bottomunit.font.family = "sans-serif", bottomunit.font.color = "black",
  bottomunit.font.size = 12, width = NULL, height = NULL)
}
\arguments{
\item{arrow.style}{character. Sets the style of the arrow. Possible values are c("triangle", "arrow", "forward arrow", "plus", "none"). "none" produces no arrow.}

\item{arrow.up}{logical. Products an upwards arrow, otherwise downwards. Result changes based on \code{arrow.style}. Defaults to \code{TRUE}.}

\item{arrow.size}{integer. Arrow size in pixels. Defaults to 15 px.}

\item{arrow.color}{HTML color string. Sets arrow color. Defaults to \code{"black"}.}

\item{arrow.width}{numeric. The proportion of the width in which the arrow is to be drawn. Valid range between \code{c(0,1)}. Without \code{topunit} or \code{bottomunit}, default is 0.5. Otherwise defaults to 0.3.}

\item{x}{character. The value to be shown.}

\item{x.font.family}{HTML font family string. Sets the font family of \code{x}. Defaults to \code{"sans-serif"}.}

\item{x.font.size}{integer. Sets the font size of \code{x} in points. Defaults to 20 pt.}

\item{x.font.color}{HTML color string. Sets the font color of \code{x}. Defaults to \code{"black"}.}

\item{x.width}{numeric. The proportion of the width in which the number is to appear. Valid range between \code{c(0,1)}. Without \code{topunit} or \code{bottomunit}, default is 0.5. Otherwise defaults to 0.4.}

\item{unit.width}{numeric. The proportion of the width in which the units are to appear. Valid range between \code{c(0,1)}. Defaults to 0.3.}

\item{unit.height}{numeric. The proportion of the height in which \code{topunit} is to appear. Valid range between \code{c(0,1)}. Defaults to 0.5.}

\item{topunit}{character. The unit on the top right corner of \code{x}. Defaults to \code{NULL}.}

\item{topunit.font.family}{HTML font family string. Sets the font family of \code{topunit}. Defaults to \code{"sans-serif"}.}

\item{topunit.font.color}{HTML color string. Sets the font color of \code{topunit}. Defaults to \code{"black"}.}

\item{topunit.font.size}{integer. Sets the font size of \code{x} in points. Defaults to 12 pt.}

\item{bottomunit}{character. The unit on the top right corner of \code{topunit}. Defaults to \code{NULL}.}

\item{bottomunit.font.family}{HTML font family string. Sets the font family of \code{bottomunit}. Defaults to \code{"sans-serif"}.}

\item{bottomunit.font.color}{HTML color string. Sets the font color of \code{bottomunit}. Defaults to \code{"black"}.}

\item{bottomunit.font.size}{integer. Sets the font size of \code{bottomunit} in pixels. Defaults to 12 pt.}
}
\description{
Creates an UpDown element.
}

