% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Box.R
\name{Box}
\alias{Box}
\title{Box}
\usage{
Box(text = NULL, horizontal_align = "left", vertical_align = "top",
  wrap_text = FALSE, background_color = "Transparent",
  font_family = "sans-serif", font_size = 11, font_bold = FALSE,
  font_italic = FALSE, font_underline = FALSE, font_strikethrough = FALSE,
  border_width = 0, border_color = "Transparent", border_style = "Solid",
  width = NULL, height = NULL)
}
\description{
Creates a Box
}

