% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhtmlMetro.R
\name{Box}
\alias{Box}
\title{Box}
\usage{
Box(text = " ", text.as.html = FALSE, horizontal.align = NULL,
  vertical.align = NULL, wrap.text = TRUE,
  background.color = "Transparent", background.shape = "Rectangle",
  font.family = "sans-serif", font.size = 11, font.color = "black",
  font.bold = FALSE, font.italic = FALSE, font.underline = FALSE,
  font.strikethrough = FALSE, border.width = 0,
  border.color = "Transparent", border.style = "Solid", width = NULL,
  height = NULL)
}
\arguments{
\item{text}{character. Adding text to the box.}

\item{text.as.html}{logical. Whether to interpret \code{text} as raw html. Defaults to \code{FALSE}.}

\item{horizontal.align}{horizontal alignment of \code{text}. Possible values are c("left","center","right"). Defaults to \code{"left"}.}

\item{vertical.align}{vertical alignment of \code{text}. Possible values are \code{c("top","middle","bottom")}. Defaults to \code{"top"}.}

\item{wrap.text}{logical. Whether text wrapping is enabled. Defaults to \code{TRUE}.}

\item{background.color}{HTML color string. Sets the background color of the box. Defaults to \code{"Transparent"}.}

\item{background.shape}{character. Specifies the shape of the background with options \code{"Rectangle", "Ellipse"}, and defaults to "Rectangle".}

\item{font.family}{HTML font family string. Sets the font family of \code{text}. Defaults to \code{"sans-serif"}.}

\item{font.size}{integer. Sets the font size of \code{text} in points. Defaults to 11 pt.}

\item{font.color}{HTML color string. Sets the font color of \code{text}. Defaults to \code{"black"}.}

\item{font.bold}{logical. Whether to bold \code{text}. Defaults to \code{FALSE}.}

\item{font.italic}{logical. Whether to have italic \code{text}. Defaults to \code{FALSE}.}

\item{font.underline}{logical. Whether to underline \code{text}. Defaults to \code{FALSE}.}

\item{font.strikethrough}{logical. Whether to draw a horizontal line across \code{text}. Defaults to code{FALSE}.}

\item{border.width}{integer. Sets the border width in pixels. Defaults to 0.}

\item{border.color}{HTML color string. Sets the border color. Defaults to "Transparent".}

\item{border.style}{HTML border style string. Sets border style. Defaults to "Solid".}

\item{width}{integer. Width of the widget in pixels. Defaults to NULL, which automatically calculates the width based on window size.}

\item{height}{integer. Height of the widget in pixels. Defaults to NULL, which automatically calculates the height based on window size.}
}
\description{
Creates an html Div container box with multiple features.
}
